requirements.txt
apscheduler
playwright
python-dotenv
requests
rich
pytest
flake8
